use mo_ctl task method to run task
